# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API for Text en/decryption requests",
        "title": "DevIntApp",
        "contact": {},
        "version": "1.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/order_text/{Id}": {
            "delete": {
                "description": "Удалить текст из запроса на шифрование по ID запроса и TextId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "text_and_order"
                ],
                "summary": "Удалить текст из запроса на шифрование",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID запроса на шифрование",
                        "name": "Id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Id текста",
                        "name": "TextId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Text was deleted from order",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schemas.ResponseMessage": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}